User_management_api/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
gunicorn==21.2.0
databases[postgresql]==0.7.0
asyncpg==0.28.0
aioredis==2.0.1
slowapi==0.1.8
orjson==3.9.7
pydantic==2.4.2
locust==2.16.1
psycopg2-binary
redis
sqlalchemy
slowapi

python-gzip-middleware
